+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+<a name="5.4.11"></a>
+## [5.4.11](https://git.softwaregroup-bg.com/ut5/ut-identity/compare/v5.4.10...v5.4.11) (2016-08-04)
+
+
+### Bug Fixes
+
+* MCTUNE-339 ([3ba5ba8](https://git.softwaregroup-bg.com/ut5/ut-identity/commit/3ba5ba8))
+
+
+
+<a name="5.4.10"></a>
+## [5.4.10](https://git.softwaregroup-bg.com/ut5/ut-identity/compare/v5.4.9...v5.4.10) (2016-08-04)
+
+
+### Bug Fixes
+
+* **identity.add:** Revert changes, as they should only appear in identity.registerRequest. ([26601db](https://git.softwaregroup-bg.com/ut5/ut-identity/commit/26601db))
+* **self-registration:** Make self registration use template for SMS. ([1bb54ad](https://git.softwaregroup-bg.com/ut5/ut-identity/commit/1bb54ad))
+
+
+
+<a name="5.4.9"></a>
+## [5.4.9](https://git.softwaregroup-bg.com/ut5/ut-identity/compare/v5.4.8...v5.4.9) (2016-08-02)
+
+
+### Bug Fixes
+
+* **self-registartion:** Make registration appear after OTP. ([327f72a](https://git.softwaregroup-bg.com/ut5/ut-identity/commit/327f72a))
+
+
+
 <a name="5.4.8"></a>
 ## [5.4.8](https://git.softwaregroup-bg.com/ut5/ut-identity/compare/v5.4.7...v5.4.8) (2016-07-25)